Finally fix the segment area handling - segments that are areas are discarded in preference to those that are not (as it was between r914 and r1136) and segments that are areas don't have the wrong distance (as they did between r914 and r1136).
Revision r1137 correctly changed to use a flag and fixed the distance bug but
then didn't sort using the new flag.  Revision r1153 started sorting using the
segment flags but the area was not the most significant bit so they were not
sorted last.  Revision r1164 correctly cleared the area flag when no longer
needed but didn't fix the rest.  Revision r1168 reverted r1164 so needed to be
re-applied.


git-svn-id: http://routino.org/svn/trunk@1169 cdd41072-61d7-4102-955e-fb406ba21da9